@@ -10,12 +10,13 @@
         }
     ],
     "require": {
-        "php": "^7.2"
+        "php": "^8.1"
     },
     "require-dev": {
-        "codeception/codeception": "^2.4 || ^3.0",
-        "spatie/phpunit-snapshot-assertions": "^1.3",
-        "digitickets/phpunit-errorhandler": "^4.0"
+        "codeception/codeception": "^5.0",
+        "spatie/phpunit-snapshot-assertions": "^4.2",
+        "digitickets/phpunit-errorhandler": "^4.1",
+        "codeception/module-asserts": "^3.0"
     },
     "autoload"          : {
         "psr-4" : {
